+++ NEWS.md
@@ -1,5 +1,9 @@
-mes v0.1.0 (Release data: 2019-12-23)
+mes v0.1.0 (Release data: 2020-03-14)
 ==============
 
 Changes:
 * Starting the package with a general description of the function. Nothing is working yet.
+* Estimation is working.
+* Occurrence model is on.
+* Forecasts with prediction intervals.
+* Model selection is now done.